kernel: ar83xx: fix igmp_snooping issue - take 2
Reverted previous fix as it is incorrect.

When IGMP snooping is enabled for the switch, CPU need to be aware
of IGMP JOIN/LEAVE as it'll be passed up the Linux stack(?).  If the
Linux bridge/router is not aware, it will likely not get upstream
router to send multicast traffic.

Tested on Askey RT4230W Rev. 6.